+++ src/api/Chat.api.ts
@@ -5,7 +5,7 @@ import { useRecoilState } from 'recoil';
 
 import { ChatDataType } from '../components/chat/ChatModal.type';
 import { CreateOrGetChatTypes, SendDirectMessageTypes } from './Api.type';
-import { ChatId } from '../atom/product.atom';
+import { ChatId } from '../atom/Product.atom';
 
 // 챗방 생성 또는 가져오기 로직을 커스텀 훅으로 분리
 export function useCreateOrGetChat() {
