+++ src/atom/Product.type.ts
@@ -1,4 +1,4 @@
-import { UnreadCount } from "./product.atom";
+import { UnreadCount } from "./Product.atom";
